Added catalog promo group aware trait to Product entity, tweaks in some twig templates and added virtual property to product for getting discount percentage.
