+++ src/Entity/Product.php
@@ -4,29 +4,30 @@ declare(strict_types=1);
 
 namespace Locastic\SyliusCatalogPromotionPlugin\Entity;
 
+use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\Product as BaseProduct;
+use Sylius\Component\Core\Model\ProductVariantInterface;
 
 class Product extends BaseProduct implements ProductInterface
 {
-    /** @var CatalogPromotionGroupInterface */
-    private $appliedCatalogPromotionGroup;
-
-    public function getAppliedCatalogPromotionGroup(): ?CatalogPromotionGroupInterface
-    {
-        return $this->appliedCatalogPromotionGroup;
-    }
-
-    public function setAppliedCatalogPromotionGroup(?CatalogPromotionGroupInterface $catalogPromotionGroup): void
-    {
-        $this->appliedCatalogPromotionGroup = $catalogPromotionGroup;
-
-        $this->getVariants()->map(function (ProductVariantInterface $productVariant) use ($catalogPromotionGroup) {
-            $productVariant->setCatalogPromotionGroup($catalogPromotionGroup);
-        });
-    }
+    use CatalogPromotionGroupAwareTrait;
 
     public function getAppliedCatalogPromotion(): ?CatalogPromotionInterface
     {
         return (null !== $this->getAppliedCatalogPromotionGroup()) ? $this->getAppliedCatalogPromotionGroup()->getCatalog() : null;
+    }
+
+    public function getPromotionDiscounts(ChannelInterface $channel)
+    {
+        $promotionDiscounts = [];
+
+        $promotionDiscounts = $this->getVariants()->map(function (ProductVariantInterface $productVariant) use ($channel, $promotionDiscounts) {
+            /** @var ChannelPricingInterface $channelPricing */
+            $channelPricing = $productVariant->getChannelPricingForChannel($channel);
+
+            return $channelPricing->getDiscount();
+        });
+
+        return $promotionDiscounts;
     }
 }
