--- src/Entity/ProductVariant.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Locastic\SyliusCatalogPromotionPlugin\Entity;
-
-use Sylius\Component\Core\Model\ProductVariant as BaseProductVariant;
-
-class ProductVariant extends BaseProductVariant implements ProductVariantInterface
-{
-    use CatalogPromotionGroupAwareTrait;
-}
