--- src/Resources/config/doctrine/ProductVariant.orm.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-    <mapped-superclass name="Locastic\SyliusCatalogPromotionPlugin\Entity\ProductVariant" table="sylius_product_variant">
-    </mapped-superclass>
-</doctrine-mapping>
